Remove side effects and improve tree-shaking - reorganize Lens, Prism, Optic namespaces into separate modules - the public API is the same
