+++ packages/focal/package.json
@@ -5,12 +5,6 @@
   "main": "dist/_cjs/src/index.js",
   "module": "dist/_esm2015/src/index.js",
   "types": "dist/_esm2015/src/index.d.ts",
-  "sideEffects": [
-    "./dist/_cjs/src/lens/index.js",
-    "./dist/_esm2015/src/lens/index.js",
-    "./dist/_cjs/src/lens/json.js",
-    "./dist/_esm2015/src/lens/json.js"
-  ],
   "files": [
     "dist/_cjs/src/",
     "dist/_esm2015/src/"
